feat: challengemember entity/repository 생성

+++ src/main/java/sarangbang/site/challengemember/repository/ChallengeMemberRepository.java
@@ -3,5 +3,5 @@ package sarangbang.site.challengemember.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import sarangbang.site.challengemember.entity.ChallengeMember;
 
-public interface ChallengeMemberRepository extends JpaRepository<ChallengeMember, Integer> {
+public interface ChallengeMemberRepository extends JpaRepository<ChallengeMember, Long> {
 }
